+++ routes/api.php
@@ -26,6 +26,11 @@ Route::post('chest', 'ChestController@store')->name('chest.store');
 Route::put('chest/{id}', 'ChestController@update')->name('chest.update');
 Route::delete('chest/{id}', 'ChestController@delete')->name('chest.delete');
 
+Route::post('album', 'AlbumController@store')->name('album.store');
+Route::post('album/upload', 'AlbumController@upload')->name('album.upload');
+Route::put('album/{id}', 'AlbumController@update')->name('album.update');
+Route::delete('album/{id}', 'AlbumController@delete')->name('album.delete');
+
 Route::get('share/{post_id}', 'ShareController@get')->name('share');
 Route::post('share/{post_id}', 'ShareController@store');